+++ CHANGELOG.md
@@ -6,6 +6,13 @@ documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 for recommendations on how to structure this file.
 
+## [0.0.2] - 2025-04-17
+
+- Support execution via dune
+- Launch terminal in the current working directory
+- Better interaction in the terminal using /usr/bin/expect
+- Support dark color theme in step execution
+
 ## [0.0.1] - 2025-04-03
 
 - Initial release